cloud_functions/vtt_to_json_converter/requirements.txt
google-cloud-storage>=2.0.0
functions-framework==3.*
webvtt-py>=0.4.6
cloudevents>=1.9.0